refactor directories and extract functions to reduce noise
