+++ kubetimer/handlers/__init__.py
@@ -1,9 +1,7 @@
-"""
-Handlers for KubeTimer operator.
+"""Handlers for KubeTimer operator.
 
 Exports:
 - Event handler functions for APScheduler-based deletion scheduling
-- Startup reconciliation for existing Deployments
 - Registry functions for memo configuration
 """
 
@@ -11,7 +9,6 @@ from kubetimer.handlers.deployment import (
     on_deployment_created_with_ttl,
     on_ttl_annotation_changed,
     on_deployment_deleted_with_ttl,
-    reconcile_existing_deployments,
 )
 
 from kubetimer.handlers.registry import (
@@ -22,6 +19,5 @@ __all__ = [
     "on_deployment_created_with_ttl",
     "on_ttl_annotation_changed",
     "on_deployment_deleted_with_ttl",
-    "reconcile_existing_deployments",
     "configure_memo",
 ]
