+++ kubetimer/scheduler/__init__.py
@@ -1,0 +1,13 @@
+"""APScheduler job management for KubeTimer.
+
+Exports:
+- schedule_deletion_job: Schedule a Deployment deletion at TTL expiry
+- cancel_deletion_job: Cancel a previously scheduled deletion job
+"""
+
+from kubetimer.scheduler.jobs import cancel_deletion_job, schedule_deletion_job
+
+__all__ = [
+    "schedule_deletion_job",
+    "cancel_deletion_job",
+]
